FEWS support for modifying subbasin params; RIVER_DEPTH constraints; natural lake h-Q disc. issue [BENCHMARKED]
Fixed support for modifying Subbasin parameters from FEWS .nc file (Model.cpp)
-required .nc read ParseNetCDFParamFile AFTER .rvh read (PArseInput.cpp)

Changed issue with poorly discretized stage-discharge curve in natural lakes for very deep lakes (found by Arelia@PCIC) (Reservoir.cpp)

Added RIVER_DEPTH constraint option to ControlStructures rules (ControlStructures.cpp)

Integrated minor GW changes from Leland (GroundwaterModel.cpp; GWRecharge.cpp;MFUSGpp.h; Model.cpp; ParseInput.cpp)

renamed LAMBDA terrain param to TOPMODEL_LAMBDA (Baseflow.cpp;ParsePropertyFile.cpp;Properties.h;TerrainClass.cpp)

fixed to support read of :SubBasinProperties without :Units line (ParseHRUFile.cpp)

git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@356 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/ParseFEWSRunInfo.cpp
@@ -704,7 +704,7 @@ bool ParseNetCDFParamFile(CModel*&pModel, const optStruct &Options)
 
           nc_get_vars_float(ncid,varids[j],start,count,stridep,ip);    HandleNetCDFErrors(retval); 
       
-          double pval=ip[start_time_index];
+          double pval=(double)ip[start_time_index];
           /*cout<<" * ";
           cout<<" |"<<param_str<<"| ";
           cout<<" |"<<class_str<<"| ";
